refactor(offense): extract shared confirmer helpers into offensive-http-common (#PR-B)
Extract the surface-resolution, path-template, read-only-path, response-shape,
and input-guard helpers out of `offensive-confirmer.js` into a new shared module
`mcp/lib/offensive-http-common.js`, so the forthcoming `bob_http_idor_confirm`
second-identity IDOR producer reuses the same security-reviewed logic instead of
forking it. This is the inert PR-B step of the offensive build sequence
(docs §11); it adds no new tool, no behavior change, no live-mint path.

Behavior-preserving: every moved symbol is a verbatim relocation (security-
rationale comments included). The only two parameterized surfaces are:

  * `assertNoForbiddenInputs(args, toolName, extraFields=[])` — the base 8-field
    denylist is shared; the confirmer passes "bob_http_confirm" and no extras, so
    its rejection message is byte-identical. The producer will pass its own tool
    name + its server-minted-only extra fields.
  * `auditConfirmRequest({..., toolId})` — the http-audit `tool` field and the
    stderr diagnostic now read the passed `toolId`; the confirmer passes its
    TOOL_ID at all three call sites, so the audit record is byte-identical.

The confirmer keeps its negative-only orchestration (`resolveConfirmSurface`,
`classifyDifferential`, `httpConfirm`, `syntheticResourceId`, method/oracle
normalizers) and re-exports `assertReadOnlyPath`/`normalizePathTemplate`/
`isResourceShapedResponse` so its public surface and the existing test imports are
unchanged. No require cycle (common imports only leaf-terminating modules; nothing
it imports reaches back to the offensive layer).

Deferred to PR-C (flagged): the net-new `responseIsSharedCacheable` from the §11
table is NOT added here — it is new logic with no confirmer caller and no
confirmer test, so it belongs with its consumer + the cache-discrimination tests
in the producer PR, keeping PR-B a true extraction.

Tests: new `test/offensive-http-common.test.js` (22 cases) unit-tests the moved
helpers directly, including the parameterization payoff (a different toolName +
extraFields drive the rejection) and the byte-identical back-compat message for
every base field. Full suite green: test:mcp 2508/0, check:syntax,
check:authority-inventory, test:prompts, and test:package (clean checkout) all
pass; the confirmer suite is unchanged and green.
