+++ test/mcp-test-manifest.json
@@ -154,6 +154,7 @@
   "test/technique-attempt-lead-surface.test.js",
   "test/claim-secret-evidence-bypass.test.js",
   "test/offensive-confirmer.test.js",
+  "test/offensive-http-common.test.js",
   "test/lifecycle-state-drift.test.js",
   "test/evidence-gate-lifecycle.test.js"
 ]
