+++ test/offensive-http-common.test.js
@@ -1,0 +1,344 @@
+"use strict";
+
+const test = require("node:test");
+const assert = require("node:assert/strict");
+const fs = require("node:fs");
+const os = require("node:os");
+const path = require("node:path");
+
+const {
+  rejectInvalidArguments,
+  decodePathSegments,
+  escapeRegExp,
+  capturedIdSegmentIsSafe,
+  pathTemplateMatchesEndpoint,
+  assertReadOnlyPath,
+  normalizePathTemplate,
+  findRoutedSurface,
+  candidateSurfaceEndpoints,
+  resolveSurfaceOrigins,
+  originFromState,
+  urlFromEndpoint,
+  resolveBaselineFromSurface,
+  isAuthChallenge,
+  isLoginRedirect,
+  responseLooksLikeLoginPage,
+  isResourceShapedResponse,
+  assertNoForbiddenInputs,
+  SCOPE_VALIDATION_OPTS,
+} = require("../mcp/lib/offensive-http-common.js");
+const { ERROR_CODES } = require("../mcp/lib/envelope.js");
+const { routeSurfaces } = require("../mcp/lib/surface-router.js");
+const { initSession } = require("../mcp/lib/session-state.js");
+const { readSessionStateStrict } = require("../mcp/lib/session-state-store.js");
+const { attackSurfacePath } = require("../mcp/lib/paths.js");
+const {
+  resetForTests: resetMaterializationDebounce,
+} = require("../mcp/lib/frontier-materialize-debounce.js");
+
+function withTempHome(fn) {
+  const previousHome = process.env.HOME;
+  const home = fs.mkdtempSync(path.join(os.tmpdir(), "bob-offensive-http-common-"));
+  process.env.HOME = home;
+  return Promise.resolve()
+    .then(() => fn(home))
+    .finally(() => {
+      if (previousHome === undefined) delete process.env.HOME;
+      else process.env.HOME = previousHome;
+      resetMaterializationDebounce();
+      fs.rmSync(home, { recursive: true, force: true });
+    });
+}
+
+function seedRoutedSurface(domain, surfaceId, endpoint) {
+  fs.mkdirSync(path.dirname(attackSurfacePath(domain)), { recursive: true });
+  fs.writeFileSync(attackSurfacePath(domain), `${JSON.stringify({
+    surfaces: [{
+      id: surfaceId,
+      title: "Synthetic API account surface",
+      surface_type: "web",
+      hosts: [domain],
+      endpoints: [endpoint],
+      tech_stack: ["fixture"],
+      priority: "HIGH",
+    }],
+  }, null, 2)}\n`, "utf8");
+  JSON.parse(routeSurfaces({ target_domain: domain }));
+}
+
+// Minimal response mock matching the {status, headers.get, bodyBytes} shape the
+// classifier helpers read (mirrors test/offensive-confirmer.test.js).
+function mockResponse(status, body = "", headers = {}, extra = {}) {
+  const lower = {};
+  for (const [k, v] of Object.entries(headers)) lower[String(k).toLowerCase()] = v;
+  return {
+    status,
+    headers: { get(name) { return lower[String(name).toLowerCase()] ?? null; } },
+    bodyBytes: Buffer.isBuffer(body) ? body : Buffer.from(body),
+    bodyTruncated: false,
+    ...extra,
+  };
+}
+
+// --- assertNoForbiddenInputs: the parameterization payoff (the whole point of PR-B) ---
+
+test("assertNoForbiddenInputs is parameterized: a different toolName + extraFields drive the rejection", () => {
+  // The forthcoming bob_http_idor_confirm reuses this guard with its own tool
+  // name and its own server-minted-only extra fields. Assert on the concrete
+  // message + code so the parameterization (toolName AND the field set) is proven.
+  let caught;
+  try {
+    assertNoForbiddenInputs({ canary: "x" }, "bob_http_idor_confirm", ["canary"]);
+  } catch (error) { caught = error; }
+  assert.ok(caught, "must throw on an extra field");
+  assert.equal(caught.code, ERROR_CODES.INVALID_ARGUMENTS);
+  assert.equal(
+    caught.message,
+    "bob_http_idor_confirm does not accept canary; the request is derived server-side from surface_id and path_template",
+  );
+  // a base field is still rejected even when the caller only supplies extras
+  let baseCaught;
+  try {
+    assertNoForbiddenInputs({ url: "https://x" }, "bob_http_idor_confirm", ["canary"]);
+  } catch (error) { baseCaught = error; }
+  assert.ok(baseCaught);
+  assert.equal(baseCaught.message, "bob_http_idor_confirm does not accept url; the request is derived server-side from surface_id and path_template");
+});
+
+test("assertNoForbiddenInputs back-compat: bob_http_confirm message is byte-identical for every base field", () => {
+  const base = ["url", "body", "headers", "severity", "demonstrated_severity", "finding_id", "resource_id", "id"];
+  for (const field of base) {
+    let caught;
+    try {
+      assertNoForbiddenInputs({ [field]: "x" }, "bob_http_confirm");
+    } catch (error) { caught = error; }
+    assert.ok(caught, `${field} must be rejected`);
+    assert.equal(caught.code, ERROR_CODES.INVALID_ARGUMENTS);
+    assert.equal(
+      caught.message,
+      `bob_http_confirm does not accept ${field}; the request is derived server-side from surface_id and path_template`,
+    );
+  }
+  // a clean confirmer args object does not throw
+  assert.doesNotThrow(() => assertNoForbiddenInputs(
+    { target_domain: "x.test", surface_id: "s", oracle_kind: "differential_response", path_template: "/api/accounts/{id}" },
+    "bob_http_confirm",
+  ));
+});
+
+test("assertNoForbiddenInputs guard semantics: nullish args are safe, hasOwnProperty (not truthiness) gates", () => {
+  assert.doesNotThrow(() => assertNoForbiddenInputs(null, "bob_http_confirm"));
+  assert.doesNotThrow(() => assertNoForbiddenInputs(undefined, "bob_http_confirm"));
+  assert.doesNotThrow(() => assertNoForbiddenInputs({}, "bob_http_confirm"));
+  // a present-but-empty forbidden field still throws (hasOwnProperty, not truthiness)
+  assert.throws(() => assertNoForbiddenInputs({ url: "" }, "bob_http_confirm"), /does not accept url/);
+  // extraFields defaults to [] — only the base list applies when omitted
+  assert.doesNotThrow(() => assertNoForbiddenInputs({ object_id: "x" }, "bob_http_confirm"));
+});
+
+// --- rejectInvalidArguments ---
+
+test("rejectInvalidArguments throws a ToolError(INVALID_ARGUMENTS) carrying details", () => {
+  let caught;
+  try {
+    rejectInvalidArguments("boom", { code: "x", path: "/p" });
+  } catch (error) { caught = error; }
+  assert.ok(caught);
+  assert.equal(caught.code, ERROR_CODES.INVALID_ARGUMENTS);
+  assert.equal(caught.message, "boom");
+  assert.equal(caught.details.path, "/p");
+});
+
+// --- escapeRegExp / decodePathSegments ---
+
+test("escapeRegExp escapes regex metacharacters", () => {
+  assert.equal(escapeRegExp("a.b+c"), "a\\.b\\+c");
+  assert.equal(escapeRegExp("/api/x"), "/api/x");
+});
+
+test("decodePathSegments recursively decodes each segment to a fixed point (<=8 passes)", () => {
+  assert.equal(decodePathSegments("/api/%2564elete/1"), "/api/delete/1"); // %2564 -> %64 -> d
+  assert.equal(decodePathSegments("/api/accounts/known"), "/api/accounts/known"); // no-op
+});
+
+// --- capturedIdSegmentIsSafe ---
+
+test("capturedIdSegmentIsSafe accepts clean ids (incl. dotted) and rejects separators/punctuation at any depth", () => {
+  for (const ok of ["known", "user.name", "1.2.3", "file.bin", "550e8400-e29b-41d4-a716-446655440000"]) {
+    assert.equal(capturedIdSegmentIsSafe(ok), true, `${ok} should be safe`);
+  }
+  for (const bad of [
+    "a/b", "a\\b", "a:b", "a;b", "a,b",
+    "known%2Fdelete", "known%252Fdelete", "known%2525252Fdelete", "known%25%32%46delete",
+    "known%5Cdelete", "known%3Acapture",
+  ]) {
+    assert.equal(capturedIdSegmentIsSafe(bad), false, `${bad} should be unsafe`);
+  }
+  assert.equal(capturedIdSegmentIsSafe(""), false);
+  // DOCUMENTED RESIDUAL (see capturedIdSegmentIsSafe comment): a dot-action suffix
+  // in a RECORDED id is treated SAFE because real ids legitimately contain dots.
+  // Pinned here so a future tightening is a deliberate, test-visible change.
+  assert.equal(capturedIdSegmentIsSafe("pay_123.capture"), true);
+});
+
+// --- pathTemplateMatchesEndpoint ---
+
+test("pathTemplateMatchesEndpoint matches single-{id} templates and rejects shape/segment mismatches", () => {
+  assert.equal(pathTemplateMatchesEndpoint("/api/accounts/{id}", "/api/accounts/known"), true);
+  assert.equal(pathTemplateMatchesEndpoint("/api/accounts/{id}", "/api/users/known"), false);
+  // captured segment routes to a sub-resource/action -> unsafe -> no match
+  assert.equal(pathTemplateMatchesEndpoint("/api/accounts/{id}", "/api/accounts/known%2Fdelete"), false);
+  assert.equal(pathTemplateMatchesEndpoint("/api/accounts/{id}", "/api/accounts/known:capture"), false);
+  // {id} that matched a multi-segment span is rejected by the [^/]+ capture
+  assert.equal(pathTemplateMatchesEndpoint("/api/accounts/{id}", "/api/accounts/a/b"), false);
+  // a template without exactly one {id} slot never matches
+  assert.equal(pathTemplateMatchesEndpoint("/api/accounts/{id}/{id}", "/api/accounts/a/b"), false);
+  assert.equal(pathTemplateMatchesEndpoint("/api/accounts/x", "/api/accounts/x"), false);
+});
+
+// --- assertReadOnlyPath / normalizePathTemplate relocation guards ---
+
+test("assertReadOnlyPath (from common) rejects destructive verb collections, allows noun reads", () => {
+  assert.doesNotThrow(() => assertReadOnlyPath("https://t.example.test/api/order/123"));
+  assert.throws(() => assertReadOnlyPath("https://t.example.test/api/delete/123"), /state-changing path segment/);
+  assert.throws(() => assertReadOnlyPath("https://t.example.test/api/%2564elete/123"), /state-changing path segment/);
+});
+
+test("normalizePathTemplate (from common) enforces the {id}-final-segment structural boundary", () => {
+  assert.equal(normalizePathTemplate("/api/accounts/{id}"), "/api/accounts/{id}");
+  assert.doesNotThrow(() => normalizePathTemplate("/api/accounts/{id}.json"));
+  assert.throws(() => normalizePathTemplate("/api/accounts/{id}/transfer"), /final path segment/);
+  assert.throws(() => normalizePathTemplate("/api/accounts/{id}.capture"), /final path segment/);
+  assert.throws(() => normalizePathTemplate("/api/accounts/{id}?fields=all"), /query string/);
+});
+
+// --- candidateSurfaceEndpoints ---
+
+test("candidateSurfaceEndpoints yields uri then string endpoints in order, skipping blank/non-string", () => {
+  assert.deepEqual(
+    candidateSurfaceEndpoints({ uri: "https://x/api", endpoints: ["https://x/a", "", 7, "https://x/b"] }),
+    [
+      { value: "https://x/api", field: "surface.uri" },
+      { value: "https://x/a", field: "surface.endpoints[0]" },
+      { value: "https://x/b", field: "surface.endpoints[3]" },
+    ],
+  );
+  assert.deepEqual(candidateSurfaceEndpoints({ endpoints: [] }), []);
+  assert.deepEqual(candidateSurfaceEndpoints(null), []);
+  assert.deepEqual(candidateSurfaceEndpoints({}), []);
+});
+
+// --- resolveSurfaceOrigins ---
+
+test("resolveSurfaceOrigins dedupes stateOrigin + endpoint origins + host-derived origins", () => {
+  const origins = resolveSurfaceOrigins({
+    uri: "https://api.example.test/v1",
+    endpoints: ["https://api.example.test/v1/x", "mailto:a@b.test", "::: not a url"],
+    hosts: ["edge.example.test", "https://api.example.test"],
+  }, "https://api.example.test");
+  assert.equal(origins[0], "https://api.example.test"); // stateOrigin first
+  assert.ok(origins.includes("https://edge.example.test")); // bare host, https from stateOrigin protocol
+  // dedupe: api.example.test appears via stateOrigin, uri, endpoint, and host -> once
+  assert.equal(origins.filter((o) => o === "https://api.example.test").length, 1);
+  // mailto / garbage endpoints contribute no origin
+  assert.equal(origins.some((o) => o.startsWith("mailto")), false);
+});
+
+test("resolveSurfaceOrigins derives host origins with http when stateOrigin is http", () => {
+  const origins = resolveSurfaceOrigins({ hosts: ["edge.example.test"] }, "http://api.example.test:8080");
+  assert.ok(origins.includes("http://edge.example.test"));
+});
+
+// --- isAuthChallenge / isLoginRedirect / responseLooksLikeLoginPage ---
+
+test("isAuthChallenge is true only for 401/403", () => {
+  assert.equal(isAuthChallenge({ status: 401 }), true);
+  assert.equal(isAuthChallenge({ status: 403 }), true);
+  assert.equal(isAuthChallenge({ status: 200 }), false);
+  assert.ok(!isAuthChallenge(null));
+  assert.ok(!isAuthChallenge(undefined));
+});
+
+test("isLoginRedirect matches auth-shaped 3xx Location only", () => {
+  assert.equal(isLoginRedirect(mockResponse(302, "", { location: "/login?next=/x" })), true);
+  assert.equal(isLoginRedirect(mockResponse(303, "", { location: "https://sso.example.test/" })), true);
+  assert.equal(isLoginRedirect(mockResponse(302, "", { location: "/dashboard" })), false);
+  assert.equal(isLoginRedirect(mockResponse(200, "", { location: "/login" })), false);
+  assert.equal(isLoginRedirect(null), false);
+  assert.equal(isLoginRedirect({ status: 302 }), false); // no headers.get
+});
+
+test("responseLooksLikeLoginPage needs html/text + a form + a login keyword", () => {
+  assert.equal(responseLooksLikeLoginPage(mockResponse(200, "<form><input name=password></form>", { "content-type": "text/html" })), true);
+  assert.equal(responseLooksLikeLoginPage(mockResponse(200, "<form><input name=q></form>", { "content-type": "text/html" })), false);
+  assert.equal(responseLooksLikeLoginPage(mockResponse(200, "<form>password</form>", { "content-type": "application/json" })), false);
+  assert.equal(responseLooksLikeLoginPage({ status: 200, headers: { get: () => "text/html" }, bodyBytes: "not a buffer" }), false);
+});
+
+// --- isResourceShapedResponse relocation smoke ---
+
+test("isResourceShapedResponse smoke: recognizes a genuine JSON resource, rejects soft-404 + empty collection", () => {
+  assert.equal(isResourceShapedResponse(mockResponse(200, "{\"id\":7,\"email\":\"a@b.test\"}", { "content-type": "application/json" })), true);
+  assert.equal(isResourceShapedResponse(mockResponse(200, "{\"error\":\"not found\"}", { "content-type": "application/json" })), false);
+  assert.equal(isResourceShapedResponse(mockResponse(200, "{\"items\":[],\"total\":0,\"page\":1}", { "content-type": "application/json" })), false);
+  assert.equal(isResourceShapedResponse(mockResponse(401, "{\"id\":7}", { "content-type": "application/json" })), false);
+});
+
+// --- findRoutedSurface (I/O) ---
+
+test("findRoutedSurface returns the routed surface and throws on an unknown id", () => withTempHome(() => {
+  const domain = "common-find.example.test";
+  const surfaceId = "surface:accounts";
+  JSON.parse(initSession({ target_domain: domain, target_url: `https://${domain}/` }));
+  seedRoutedSurface(domain, surfaceId, `https://${domain}/api/accounts/known`);
+
+  const { route, surface } = findRoutedSurface(domain, surfaceId);
+  assert.equal(route.surface_id, surfaceId);
+  assert.equal(surface.id, surfaceId);
+
+  let caught;
+  try { findRoutedSurface(domain, "surface:does-not-exist"); } catch (error) { caught = error; }
+  assert.ok(caught);
+  assert.equal(caught.code, ERROR_CODES.INVALID_ARGUMENTS);
+  assert.match(caught.message, /unknown or unrouted surface_id surface:does-not-exist/);
+}));
+
+// --- originFromState / urlFromEndpoint / resolveBaselineFromSurface (I/O) ---
+
+test("originFromState returns the in-scope origin and rejects a missing/invalid target_url", () => withTempHome(() => {
+  const domain = "common-origin.example.test";
+  JSON.parse(initSession({ target_domain: domain, target_url: `https://${domain}/` }));
+  assert.equal(originFromState(domain, { target_url: `https://${domain}/app` }), `https://${domain}`);
+  assert.throws(() => originFromState(domain, {}), /requires a web session with target_url/);
+  assert.throws(() => originFromState(domain, { target_url: "not a url" }), /not a valid URL/);
+}));
+
+test("urlFromEndpoint resolves absolute URLs and absolute paths, rejects bare tokens", () => {
+  assert.equal(urlFromEndpoint("https://x.test/api", "https://x.test", "f").toString(), "https://x.test/api");
+  assert.equal(urlFromEndpoint("/api/y", "https://x.test", "f").toString(), "https://x.test/api/y");
+  assert.throws(() => urlFromEndpoint("api/y", "https://x.test", "f"), /must be an absolute http\(s\) URL or an absolute path/);
+});
+
+test("resolveBaselineFromSurface resolves a query-free baseline that matches the template, else throws", () => withTempHome(() => {
+  const domain = "common-baseline.example.test";
+  JSON.parse(initSession({ target_domain: domain, target_url: `https://${domain}/` }));
+  const { state } = readSessionStateStrict(domain);
+  const surface = { id: "s", hosts: [domain], endpoints: [`https://${domain}/api/accounts/known?fields=all`] };
+
+  const baseline = resolveBaselineFromSurface({ domain, surface, pathTemplate: "/api/accounts/{id}", state });
+  // recorded query is dropped for baseline/target symmetry
+  assert.equal(baseline.toString(), `https://${domain}/api/accounts/known`);
+  assert.equal(baseline.search, "");
+
+  assert.throws(
+    () => resolveBaselineFromSurface({ domain, surface, pathTemplate: "/api/users/{id}", state }),
+    /path shape does not match any recorded endpoint for surface_id/,
+  );
+}));
+
+// --- SCOPE_VALIDATION_OPTS export contract ---
+
+test("SCOPE_VALIDATION_OPTS is the frozen intentional-no-internal-host-block sentinel", () => {
+  assert.deepEqual(SCOPE_VALIDATION_OPTS, { blockInternalHosts: false });
+  assert.equal(Object.isFrozen(SCOPE_VALIDATION_OPTS), true);
+});
